hifi4: fix race condition issue for shm alloc [1/1]
PD#SWPL-16577

Problem:
We are using fd as the flag, representing the initialization
of all global shared memory context are ready. This is wrong.
Since the initialization of rpc handle is later than fd.

Solution:
The best solution is using a dedicate flag to represent init done.
This flag is set at the end of initialization. So far, we use
rpc handle as the flag, since the rpc handle are the latest context
initialized.

Verify:
A1

Change-Id: I7b15429deb8de3e65ff0234fef33554a9b163282
Signed-off-by: Wenjie Zhou <wenjie.zhou@amlogic.com>